requirements.txt
python-telegram-bot==20.7
telethon==1.33.1
SQLAlchemy==2.0.25
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
flask==3.0.1
requests==2.31.0 